+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Postman Collection SDK Changelog
 
+* Enhanced the `PropertyList` to allow keys with multiple values
+* Removed Response details `fromServer` flag
+* Added Response details `standardName` property
+* Added `set`, `get`, `unset` and `clear `helpers to `VariableScope`
+* Fixed a bug in `PropertyList#clear` which caused only the first property of a list to be removed
+* Added `Response#details` helper to retain custom server HTTP reasons
+* Fixed a script resolution bug in `Event#update`
+* Added `Response.isResponse` and `Script.isScript` helpers
+* Added `Request.isRequest` helper
+
 #### 1.0.2 (March 21, 2017)
 * Downgraded `file-type` to v3.9.0, which is browserify compatible